pyproject.toml
[project]
name = "ai-capabilities-explorer"
version = "0.1.0"
description = "Gemma 4 + SAM 3.1 local vision-guided segmentation on Apple Silicon"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "mlx-vlm>=0.4.3",
    "numpy>=2.0.0",
    "pillow>=12.2.0",
    "python-multipart>=0.0.22",
    "uvicorn>=0.43.0",
]
